backend/requirements.txt
Flask
Flask-Cors
spacy
selenium
openai
docx2txt
transformers
torch
python-docx
PyMuPDF
pdfminer.six
sentencepiece
python-dotenv
requests
google-generativeai